5411: pagination hook
