+++ src/core/configuration/dpl-config.json
@@ -1,9 +1,9 @@
 {
   "mobile": {
-    "pageSize": 20
+    "pageSize": 2
   },
   "desktop": {
-    "pageSize": 50
+    "pageSize": 5
   },
   "coverTints": ["100", "40", "80", "120", "20"]
 }
