--- src/core/configuration/pagesize.json
@@ -1,3 +1,0 @@
-{
-  "pageSize": 25
-}
